refactor(usage): split endpoint tests
